requirements.txt
python<4>3.8
pyyaml<6